refactor menu nav bar
This change refactors the menu bar, the main purpose for this is to make
expansion and maintenance easier. Previously the process involved much
repeated code and required a heavy duplication process for adding items.
This style has been dropped in favor a lighter version with less
duplications and only a single file to maintain for the main
radanalytics page. The tutorial applications menu has not been changed.

Changes
* add a new page configuration item for menu entry
* add a new combined menu template for radanalytics.io
* remove old menu templates: community, faq, projects, tutorials,
    welcome
* update main site pages front matter to include new entry variable
* update configuration with default values