add bower to the mix
This change adds a bower configuration file to handle the patternfly
dependency. As the usage of patternfly grows within the site, extra
files are needed to make it function properly(most notably, jquery).
Instead of managing the dependencies by hand, bower will allow us to use
automated tooling. The site is meant to be produced by github pages
jekyll, and as such, the bower_components directory is being checked in.