convert presentations to asciidoc
This change takes the presentations out of the data file they were in
and converts them to individual AsciiDoc files. The configuration has
been updated to remove the old data and add the new collection.
Additionally, the community page template has been upgraded to
accomodate a date based sorting and the new file variable format.

The contribution documentation and templates have been updated to
include this new process for adding presentations to the site.