Add integration tests for changes to resources.yaml
This change adds travis tests which check whether or not
resources.yaml in the current pull request or branch differs
from resources.yaml on the master branch. If resources.yaml
is different, the template integration tests from the oshinko-s2i
repository are run against the templates created by resources.yaml.
If resources.yaml is not different, the test returns success.