add the My First RADanalytics application
This tutorial contains a top level page and a few sub-level pages. The
sub-levels are contained in the assets folder for the top level
namespace.

This change also makes a few modifications to the page layouts to help
with the spacing at the bottom of pages and the background of code
blocks.

An entry is added to the front page for this new tutorial and it is
removed from the application list.