update oshinko webui cluster launch article
This change refines the language in the "how do i" article for deploying
spark clusters from oshinko webui. It adds the modular doc template and
incorporates suggestions on style and grammar.

The advanced option has been removed as it will be proposed as a
separate "how do i" article.

changes
* add modular doc template to deploy-a-spark-cluster-webui.adoc
* many grammar and style fixes
* trimmed images to not leave open space
* removed advanced options from deployment doc