fix titles in how do i articles
Previously, the how do i articles all assumed that a writer would create
a title that did not include the "How do I" preamble. This resulted in
document titles that looked incomplete and poorly formed. This change
makes it so that the individual article titles must include the full
sentence including "How do I" at the beginning.